Create Srv_Connectivity_Test.ps1
This script will first verify the connectivity to each server in the serverList array. If the server is online, it will then check the specified portRange to find open ports. Finally, it will output the open ports in a file named OpenPorts_<timestamp>.txt in the C:\temp directory. The user running the script needs to have the necessary permissions to access the target servers and write to the C:\temp directory.